requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
chromadb==0.4.15
sentence-transformers==2.2.2
PyMuPDF==1.23.8
pdfminer.six==20221105
google-generativeai==0.3.2
python-dotenv==1.0.0
langchain==0.0.335
langchain-google-genai==0.0.6
numpy==1.24.3
pandas==2.0.3
pydantic==2.5.0
httpx==0.25.2
loguru==0.7.2